refactoring into CardGrid
