+++ src/styles/App.css
@@ -21,12 +21,9 @@ body {
   left: 0;
   right: 0;
   top: 0;
-  background: rgb(0, 0, 0);
-  background: rgba(0, 0, 0, 0.5);
-  color: #f1f1f1;
+  background: #00000080;
   transition: 0.5s ease;
   opacity: 0;
-  color: white;
   font-size: 20px;
   padding: 20px;
   text-align: center;
@@ -40,5 +37,5 @@ body {
 }
 
 .details:hover {
-  color: blue;
+  color: #0000ff;
 }
